+++ queries/query2ADT.h
@@ -23,6 +23,6 @@ void freeQueryQ2(query2ADT q);
 
 size_t countQ2(const query2ADT q);
 
-void freeFinalVecQ2(DataQ2 * vec, int dim);
+void freeFinalVecQ2(DataQ2 * vec, size_t dim);
 
 #endif
